Remove simple warnings

+++ Source/MediaStorageAndFileFormat/gdcmRLECodec.cxx
@@ -570,7 +570,7 @@ size_t RLECodec::DecodeFragment(Fragment const & frag, char *buffer, size_t llen
   is.write(mybuffer, bv.GetLength());
   delete[] mybuffer;
   std::stringstream os;
-  SetLength( llen );
+  SetLength( (unsigned long)llen );
 #if !defined(NDEBUG)
   const unsigned int * const dimensions = this->GetDimensions();
   const PixelFormat & pf = this->GetPixelFormat();
@@ -651,7 +651,7 @@ bool RLECodec::Decode(DataElement const &in, DataElement &out)
         gdcmDebugMacro( "RLE pb with frag: " << i );
         corruption = true;
       }
-      pos += llen;
+      pos += (unsigned long)llen;
       }
     if( !corruption )
       assert( pos == len );
